2 coercions to quite MSVC. Eliminate stray letter, unclear how it got there...
git-svn-id: svn+ssh://svn.code.sf.net/p/xsb/src/trunk@7853 756869e0-9e47-4bc5-b0a8-847305dc5e1c

+++ XSB/emu/system_xsb.c
@@ -197,7 +197,7 @@ int sys_syscall(CTXTdeclc int callno)
     break;
   }
   case SYS_epoch_seconds: {
-    ctop_int(CTXTc 3,time(0));
+    ctop_int(CTXTc 3,(Integer)time(0));
     break;
   }
   default: xsb_abort("[SYS_SYSCALL] Unknown system call number, %d", callno);
